+++ tests/pna/collapse/test_paired_collapse.py
@@ -37,7 +37,7 @@ def test_split_chunks():
 
 @pytest.mark.slow
 def test_collapse_from_paired_demux_data(tmp_path, testdata_paired_small_demux):
-    assay = pna_config.get_assay("pna-2")
+    assay = pna_config.get_assay("proxiome-v1")
     panel = pna_config.get_panel("proxiome-immuno-155-v3")
     output = tmp_path / "test.parquet"