+++ tests/pna/demux/test_barcode_identifier.py
@@ -18,7 +18,7 @@ from pixelator.pna.demux.barcode_identifier import BarcodeIdentifier
 
 
 def test_barcode_identifier_exact_match():
-    assay = pna_config.get_assay("pna-2")
+    assay = pna_config.get_assay("proxiome-v1")
     panel = pna_config.get_panel("proxiome-immuno-155-v3")
 
     barcodes_id = BarcodeIdentifier(assay=assay, panel=panel, mismatches=1)
@@ -50,7 +50,7 @@ def test_barcode_identifier_exact_match():
 
 
 def test_barcode_identifier_one_mismatch():
-    assay = pna_config.get_assay("pna-2")
+    assay = pna_config.get_assay("proxiome-v1")
     panel = pna_config.get_panel("proxiome-immuno-155-v3")
 
     barcodes_id = BarcodeIdentifier(assay=assay, panel=panel, mismatches=1)
@@ -82,7 +82,7 @@ def test_barcode_identifier_one_mismatch():
 
 
 def test_barcode_identifier_failed():
-    assay = pna_config.get_assay("pna-2")
+    assay = pna_config.get_assay("proxiome-v1")
     panel = pna_config.get_panel("proxiome-immuno-155-v3")
 
     with tempfile.NamedTemporaryFile() as buffer: