+++ CHANGELOG.md
@@ -90,7 +90,7 @@ which kit lot has been used to produce the data. See the table below for which p
 
 - Panel data can now be fetched from the `.pxl` file in the denoise step
 - `control` is now stored as boolean in the Panel object, instead of `yes`/`no`.
-- Fix an error in the "Linker binding site 1" region of the `pna-2` assay config file
+- Fix an error in the "Linker binding site 1" region of the `proxiome-v1` assay config file
 
 ## [0.22.0] - 2025-10-23